Add Dockerfile, docker-compose.yml, and update composer.json
This commit adds the Dockerfile and docker-compose.yml files for containerization. It also updates the composer.json file by adding a version number. Additionally, a new index.php file is created to display the number of items returned by the Laravel-News.com feed. Lastly, the namespace in the Api.php file is corrected from "LarevelNews" to "LaravelNews".